uploading

+++ internal/domain/coin/model/model.go
@@ -2,8 +2,8 @@ package model
 
 type Coin struct {
 	Name            string  `json:"name"`
-	Price           int     `json:"price"`
-	MinPrice        int     `json:"min_price"`
-	MaxPrice        int     `json:"max_price"`
+	Price           float64 `json:"price"`
+	MinPrice        float64 `json:"min_price"`
+	MaxPrice        float64 `json:"max_price"`
 	HourChangePrice float64 `json:"hour_change_price"`
 }
